exercises/qa/fastapi/pyproject.toml
[tool.poetry]
name = "fastapi-exercise"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
pydantic = "^2.7.1"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.2"
requests = "^2.32.3"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
